implemented mod menu config added frost green color preset fixed blinking UI elements

+++ src/main/java/net/spellcraftgaming/rpghud/gui/hud/element/modern/HudElementArmorModern.java
@@ -38,7 +38,7 @@ public class HudElementArmorModern extends HudElement {
 	
     @Override
     public int getPosX(int scaledWidth) {
-        return (int) ((scaledWidth / 2 - 91)*getInvertedScale() + this.settings.getPositionValue(Settings.armor_position)[0]);
+        return (int) (((float) scaledWidth / 2 - 91)*getInvertedScale() + this.settings.getPositionValue(Settings.armor_position)[0]);
     }
 
     @Override
